+++ tests/providers/stackit/services/iaas/iaas_security_group_all_traffic_unrestricted/__init__.py
@@ -1,1 +1,0 @@
-# StackIT IaaS Security Group All Traffic Unrestricted Check Tests