+++ tests/providers/stackit/services/iaas/iaas_security_group_database_unrestricted/__init__.py
@@ -1,1 +1,0 @@
-# StackIT IaaS Security Group Database Unrestricted Check Tests