+++ tests/providers/stackit/services/iaas/iaas_security_group_rdp_unrestricted/__init__.py
@@ -1,1 +1,0 @@
-# StackIT IaaS Security Group RDP Unrestricted Check Tests